Split work-target availability API into own file for repo line limit.
order_suggestion_work.dart exceeded repo.dart_file_non_comment_line_size (1000).
Move AvailableWorkTargetsForUnit and getAvailableWorkTargetsForUnit to
order_suggestion_work_availability.dart; re-export from order_suggestion.dart.

Refs #2133.
