+++ packages/colonizethis_logic/lib/src/orders/order_suggestion.dart
@@ -44,8 +44,6 @@ export 'order_suggestion_naval_diplomatic.dart'
         suggestDiplomaticOrders,
         suggestNavalMissionOrders,
         suggestNavalMoveOrders;
-export 'order_suggestion_work.dart'
-    show
-        AvailableWorkTargetsForUnit,
-        getAvailableWorkTargetsForUnit,
-        suggestWorkOrders;
+export 'order_suggestion_work.dart' show suggestWorkOrders;
+export 'order_suggestion_work_availability.dart'
+    show AvailableWorkTargetsForUnit, getAvailableWorkTargetsForUnit;
